% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_functions.R
\name{ccmpp_c}
\alias{ccmpp_c}
\title{\R wrapper for the _C_ function \code{C_ccmpp}.}
\usage{
ccmpp_c(
  out_pop_fem,
  out_pop_male,
  srb,
  fert,
  surv_fem,
  surv_male,
  surv_fem_0,
  surv_male_0,
  mig_fem,
  mig_male,
  bline_fem,
  bline_male,
  n_proj_steps,
  n_age_pop,
  step_wid
)
}
\arguments{
\item{out_pop_fem}{(\code{double *out_pop_fem}) In \R this is an
\code{n_age_pop} * (1 + \code{proj_steps}) matrix with age
groups as rows and time periods as columns. In _C_ it is an
array of length \code{n_age_pop} * (1 +
\code{proj_steps}). Its first column (in \R), or first
\code{n_age_pop} entries (in _C_) contain \code{bline_fem},
the age-specific female population counts in the year that
marks the start of the projection (the \dQuote{baseline}). The
rest of the entries are arbitrary; they serve as a repository
for the results which will overrite whatever entries are
supplied. See \dQuote{Details}.}

\item{out_pop_male}{(\code{double *out_pop_male}) Same as
\code{out_pop_fem} but for males.}

\item{srb}{(\code{double *srb}) Vector of length
\code{n_proj_steps} containing the average sex ratios at
birth.}

\item{fert}{(\code{double *fert}) Average annual age-specific
fertility rates for each of the \code{n_proj_steps}. In \R
this can be supplied as a matrix with age groups as rows and
time periods as columns. In _C_ it is an array of length
\code{n_age_pop} * \code{proj_steps} formed by concatenating
the age-specific fertility rates for each projection period.}

\item{surv_fem}{(\code{double *surv_fem}) Age-specific survival
proportions for females for ages 5 to the open ended age
group. Like \code{fert}, in \R it can be supplied as an
\code{n_age_pop} by \code{n_proj_steps} matrix with age groups
as rows. If using the _C_ version it should be an array with
\code{n_age_pop} * \code{n_proj_steps} elements.}

\item{surv_male}{(\code{double *surv_male}) Same as
\code{surv_fem} but for males.}

\item{surv_fem_0}{(\code{double *surv_fem_0}) Vector of lenght
\code{n_proj_steps} containing survival proportions of females
for the period birth to the end of the first projection
interval, for each period in the projection.}

\item{surv_male_0}{(\code{double *surv_male_0}) Same as
\code{surv_fem_0} but for males.}

\item{mig_fem}{(\code{double *mig_fem}) A matrix (in \R) or vector
(in _C_) with the same dimensions as for \code{fert} but
containing average annual net migrations for females as a
proportion of the receiving population, by age.}

\item{mig_male}{(\code{double *mig_male}) Same as \code{mig_fem}
but for males.}

\item{bline_fem}{(\code{double *bline_fem}) A vector of
age-specific female population counts in the year that marks
the start of the projection (the \dQuote{baseline}).}

\item{bline_male}{(\code{double *bline_male}) Same as
\code{bline_fem} but for males.}

\item{n_proj_steps}{(\code{int *n_proj_steps}) The number of steps
to project forward. This must conform to the dimensions of the
other inputs; see where this argument is mentioned in their
descriptions. \emph{No checking is done} to make sure this
condition is satisfied.}

\item{n_age_pop}{(\code{int *n_age_pop}) The number of age groups
in the projection. As for \code{n_proj_steps}, this must
conform to the dimensions of the other inputs; see where this
argument is mentioned in their descriptions. \emph{No checking
is done} to make sure this condition is satisfied.}

\item{step_wid}{(\code{int *step_wid}) The scale of the projection
intervals, e.g., in years.}
}
\value{
In \R, a two-element list containig the first two
    arguments with their (\code{n_age_pop} + 1), (\code{n_age_pop}
    + 2), ..., elements replaced with the projected counts for
    females and males, respectiely. The _C_ function returns
    nothing (\code{void}) but has the side effect of similarly
    replacing element in the first two arguments.
}
\description{
This is a simple wrapper for the _C_ function of the same name
which performs population projection by the cohort component
method (see the documentation for \code{\link{ccmpp}}). The _C_
function can be loaded for use in _C_ source code via
\code{#include ccmppAPI}. The arguments are documented here for
both functions. Argument types for _C_ are given in parentheses in
the argument list.
}
\details{
The _C_ code is accessed via \code{\link{.C}} hence all arguments
are passed via pointers. In particular, the first two arguments
are containers for the output. They must be vectors of length
\code{n_age_pop} which can be coerced via
\code{\link{as.double}}. The values of the elements are
unimportant as they will be overwritten with the output; e.g., you
could supply the same value to \code{out_pop_fem} as
\code{bline_fem} (and similarly for the male versions), as in the
example.
}
\examples{
data("Thailand_demog")
raw <-
  with(Thailand_demog, {
  ccmpp_c(out_pop_fem = matrix(rep(thai_base_pop_counts$female, 9), ncol = 9),
          out_pop_male = matrix(rep(thai_base_pop_counts$male, 9), ncol = 9),
          srb = thai_srb,
          fert = thai_fert_rates,
          surv_fem = thai_surv_props$female[-1,],
          surv_male = thai_surv_props$male[-1,],
          surv_fem_0 = thai_surv_props$female[1,],
          surv_male_0 = thai_surv_props$male[1,],
          mig_fem = thai_mig_props$female,
          mig_male = thai_mig_props$male,
          bline_fem = thai_base_pop_counts$female,
          bline_male = thai_base_pop_counts$male,
          n_proj_steps = ncol(thai_fert_rates), #8
          n_age_pop = nrow(thai_base_pop_counts$female), #17
          step_wid = 5
  )})

(nicer <- list(female = matrix(raw[[1]], ncol = 9),
              male = matrix(raw[[2]], ncol = 9)))

}
\seealso{
\code{\link{ccmpp}}, a more friendly wrapper for
    the same underlying _C_ function that takes arguments as \R
    lists and does some optional input checks.
}
\author{
Mark C Wheldon
}
