% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_functions.R
\name{ccmpp_c}
\alias{ccmpp_c}
\title{Cohort-component population projection (C implementation)}
\usage{
ccmpp_c(
  base_pop_counts,
  surv_props,
  fert_rates,
  srb = matrix(1.05, ncol = proj_steps),
  mig_props,
  proj_steps = ncol(fert_rates),
  n_age_grps = length(base_pop_counts$female),
  age_int = 5,
  label_dims = FALSE,
  base_year = colnames(base_pop_counts[[1]]),
  first_age = 0,
  do_checks = FALSE,
  verbose = FALSE,
  return_list = FALSE
)
}
\arguments{
\item{base_pop_counts}{Population counts at baseline. List with up to two components: \dQuote{female} and \dQuote{male} each a column vector of age specific counts.}

\item{surv_props}{List with up to two components: \dQuote{female} and \dQuote{male} each a matrix of survivorship proportions: the probability of reaching the age at the start of the interval for each projection interval. Years as columns, ages as rows. The first row should be nL0/(age_int*l0). The last row is survival for age_int years in the open interval.}

\item{fert_rates}{Matrix of (annual) age-specific fertility rates (NOT yet multiplied by \code{age_int}) for each projection interval. Years as columns, ages as rows.}

\item{srb}{Matrix of sex ratios at birth. Should be of dim 1 * \code{proj_steps}.}

\item{mig_props}{List with up to two components: \dQuote{female} and \dQuote{male} each a matrix of age-specific proportions for each projection interval, beginning with baseline. Years as columns, ages as rows.}

\item{proj_steps}{The number of time periods to project forward.}

\item{n_age_grps}{Number of age groups}

\item{age_int}{Size of projection intervals (years).}

\item{label_dims}{Should output have dimnames set? (cosmetic).}

\item{base_year}{Label of baseline year (cosmetic).}

\item{first_age}{Label of first age group (cosmetic).}

\item{do_checks}{Logical; check inputs for validity?}

\item{verbose}{Logical; issue messages (cosmetic)?}

\item{return_list}{Logical. If female only projection, should a list be returned with one component called \dQuote{female}?}
}
\value{
If \code{isTRUE(return_list)}, a list with up to two components, \dQuote{female} and \dQuote{male}, each a matrix of age-specific population counts, years as columns, ages as rows. Otherwise a matrix of age-specific counts for females only.
}
\description{
A C implementation of the two-sex, female-dominant cohort-component method of
population projection (e.g., \cite{Preston et. al, 2001, Ch. 6}). This is a deterministic method for
projecting age-stratified population counts forward in time.
}
\examples{
data("Thailand_demog")
with(Thailand_demog, ccmpp_c(thai_base_pop_counts,
                  surv_props = thai_surv_props,
                  fert_rates=thai_fert_rates,
                  srb = thai_srb, mig_props = thai_mig_props))
}
\references{
Preston, S. H., Heuveline, P., and Guillot, M. (2001), \emph{Demography: Measuring and Modeling Population Processes}, Malden, Massachusetts: Blackwell.
}
\seealso{
Other CCMPP backend functions: 
\code{\link{ccmpp_r}()}
}
\author{
Mark Wheldon
}
\concept{CCMPP backend functions}
